backend/services/identity-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
pydantic==2.11.9
pydantic-settings==2.10.1
PyJWT==2.10.1
bcrypt==4.3.0
email-validator==2.3.0
psycopg[binary]==3.2.10
httpx==0.28.1
python-json-logger==3.3.0
